Add views for fps and dropped frames.

+++ core/src/main/java/com/novoda/noplayer/NoPlayer.java
@@ -257,7 +257,7 @@ public interface NoPlayer extends PlayerState {
 
     interface FramesPerSecondChangedListener {
 
-        void onFramesPerSecondChanged(double framesPerSecond);
+        void onFramesPerSecondChanged(int framesPerSecond);
     }
 
     /**
